# Specialize `list_value` for primitive types for significantly improved performance