--- a/.github/regression/micro_extended.csv
+++ b/.github/regression/micro_extended.csv
@@ -164,6 +164,7 @@ benchmark/micro/list/column_data_collection_copy/unnest.benchmark
 benchmark/micro/list/list_few_rows.benchmark
 benchmark/micro/list/list_many_rows.benchmark
 benchmark/micro/list/list_order_by.benchmark
+benchmark/micro/list/list_value.benchmark
 benchmark/micro/list/string_split.benchmark
 benchmark/micro/list/string_split_regexp.benchmark
 benchmark/micro/list/string_split_unicode.benchmark