--- a/benchmark/micro/list/list_order_by.benchmark
+++ b/benchmark/micro/list/list_order_by.benchmark
@@ -2,7 +2,7 @@
 # description: Ordered LIST aggregation
 # group: [list]
 
-name String Split Regexp
+name List Order By
 group micro
 subgroup list