This PR specializes `list_value` for primitive types, greatly speeding it up over the `Value` based method used previously. For example:

```sql
create table integers as from range(10000000) t(i);
create table lists as select [i] from integers;
```

| v1.0.0 |  New  |
|--------|-------|
| 0.94s  | 0.15s |

#### Longer Lists

```sql
create table lists as select [i, i + 1, i + 2, i + 3] from integers;
```

| v1.0.0 |  New  |
|--------|-------|
| 3.63s  | 0.33s |